Secondary particle kinematics data-mc comparisons

+++ Helpers.h
@@ -161,7 +161,7 @@ double PION_SCATTERING_ENERGY_THRESHOLD = 0.05; // in GeV / c
 
 // Bins for energy bins
 static const std::vector<double> ARRAY_KE_BINS = {0., 100., 200., 300., 400., 600.};
-static const int                   NUM_BINS_KE = ARRAY_KE_BINS.size() - 1;
+static const int                   NUM_BINS_KE = static_cast<int>(ARRAY_KE_BINS.size()) - 1;
 
 // TOF mass cut
 double PI_MU_EL_MASS_CUTOFF = 350.;
